# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  

  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:5000/api
...
paths:
  /user/register:
    post:
      tags:
        - user
      security: []
      requestBody:
        description: user info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      responses:
        '200':
          description: Successful 
        '500':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
  /user/login:
    post:
      tags:
        - user
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      description: get jwt token
      responses:
        '200':
          description: Successful 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '401':
          description: username or password wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
        '500':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
  /user/refresh:
    post:
      tags:
        - user
      description: refresh jwt token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
        '500':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
...
components:
  schemas:
    BaseData:
      type: object
      properties:
        msg:
          type: string
          example: "system msg"
    ErrorMsg:
      allOf:
        - $ref: '#/components/schemas/BaseData'
    JwtToken:
      allOf:
        - $ref: '#/components/schemas/BaseData'
        - type: object
          properties:
            token:
              type: string
            refresh_token:
              type: string
    UserInfo:
      type: object
      properties:
        username:
          type: string
          example: "user1"
        password:
          type: string
          example: "123456"
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 